CRITICAL SECURITY UPDATE
Fixed a bug with website authentication.

This bug allowed both logged in and unauthenticated users
to perform actions that should not be possible by editing
a number of hidden fields. This included editing posts and
granting privileges to users.

The contents of these hidden fields are now validated server-side,
preventing this attack.

This fix also prevents users from circumventing authentication
by going to specific URLs directly.